APIS-4714 - Moving more tests
