+++ test/uk/gov/hmrc/apidocumentation/controllers/ApiDocumentationControllerSpec.scala
@@ -18,14 +18,13 @@ package uk.gov.hmrc.apidocumentation.controllers
 
 import play.api.mvc.MessagesControllerComponents
 import uk.gov.hmrc.apidocumentation.models._
-import uk.gov.hmrc.apidocumentation.services.DocumentationService
 import uk.gov.hmrc.apidocumentation.views.html._
 import uk.gov.hmrc.apidocumentation.ErrorHandler
 import play.api.http.Status.{INTERNAL_SERVER_ERROR, OK, SEE_OTHER, NOT_FOUND}
 import uk.gov.hmrc.apidocumentation.utils.ApiDefinitionTestDataHelper
-import org.mockito.Mockito.{when,verify}
+import org.mockito.Mockito.when
 import org.mockito.Matchers.any
-import uk.gov.hmrc.apidocumentation.services.{PartialsService, RAML}
+import uk.gov.hmrc.apidocumentation.services.RAML
 import uk.gov.hmrc.ramltools.domain.{RamlParseException, RamlNotFoundException}
 import scala.concurrent.Future.failed
 import uk.gov.hmrc.http.NotFoundException
@@ -33,8 +32,8 @@ import uk.gov.hmrc.apidocumentation.controllers.utils._
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class ApiDocumentationControllerSpec extends CommonControllerBaseSpec with PageRenderVerification with ApiDefinitionTestDataHelper with RamlPreviewBaseSpec {
-  trait Setup extends ApiDocumentationServiceMock {
+class ApiDocumentationControllerSpec extends CommonControllerBaseSpec with PageRenderVerification with ApiDefinitionTestDataHelper {
+  trait Setup extends ApiDocumentationServiceMock with AppConfigMock with ApiDefinitionServiceMock with LoggedInUserProviderMock {
     val errorHandler = app.injector.instanceOf[ErrorHandler]
     val mcc = app.injector.instanceOf[MessagesControllerComponents]
